requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# AWS/Cloud storage
boto3>=1.28.0

# Configuration and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0
PyYAML>=6.0

# Data profiling and validation
great-expectations>=0.17.0
ydata-profiling>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Logging and monitoring
structlog>=23.1.0

# Development tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter support
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Optional: Spark support (install separately if needed)
# pyspark>=3.4.0
# findspark>=2.0.1